Don't waste space when inplace_function is overaligned for its size.
Arguably this is a pessimization and we ought to place the most
commonly used member of the object (`m_ManagerFctPtr`) at offset
zero instead.  Do we care?  Maybe not.

This change affects layout only when `Alignment > Capacity && Alignment > 16`.
At all other times it's just shuffling around the *locations* of the
padding bytes, not affecting the *number* of padding bytes.