inplace_function: "Optimize" swap for trivially relocatable types.
This patch implements the purported-optimization described in this
forum thread:
https://groups.google.com/a/isocpp.org/d/msg/std-proposals/HGCHVSRwSMk/k7Ir-rmxBgAJ

In the old code, `f.swap(g)` always performs six indirect calls.

In the new code, `f.swap(g)` first performs one or two indirect loads,
and if both results are "okay", it just swaps the byte-level representations.
If either result is "not okay", then it performs the six indirect calls.

This is probably a speedup in the "okay" case, but I'd have to see
some benchmarks.