[aa_inplace_vector] Copying/moving an Alloc is not allowed to throw
The Standard specifies in [allocator.requirements.general] that copy
and move operations on allocators never throw, regardless of whether
they're marked noexcept or not. And in fact `boost::interprocess::allocator`
does not mark its SMFs as noexcept. So we shouldn't check the trait
here; instead we should assume that the operations are non-throwing.